Do not create additional directory 'eclipse' in droplet folder.
Eclipse dropins supported being present directly in the dropin folder,
or under an additonally placed 'eclipse' folder. During migration to
droplets we decided to honour this format but additional patching is
needed in the platfrom to check the extra location as droplet logic
expects the content to be directly under the droplet folder.